ci(Update): Add SAVBLO button and checkbox to PlaningChaud
A new button and checkbox related to SAVBLO activity are added in
PlaningChaud.xaml. The corresponding click event handler is also created
in PlaningChaud.xaml.cs. The export validation logic is modified to
reflect this addition by checking if all the checkboxes are selected
instead of a fixed number. The SAVBLO activity is added to EActivity.cs.